Add files via upload
This calculator project was created using Python and provides basic arithmetic operations such as addition, subtraction, multiplication, division; it can also find the remainder and power(Exponential Function). The code is designed to be easily customizable and expandable, making it a great starting point for learning more advanced programming concepts.