Add files via upload
"Faulty Calculator" is a unique project that may seem like a regular calculator at first glance, but it has a fun twist. It is designed to perform all the calculations correctly, but with a deliberate flaw - it gives incorrect answers for certain specific values. This project aims to challenge and entertain users while also improving their calculation skills.